Finished up basic room implementation
Added 4 Scripts which are as follows:

Floor: Handles all the rooms and player position.

Camera Controller: Handles all of the camera movement

Room: Handles all room related things (which door was walked through, etc)

RoomDoor: Handles collision with doors

In the floor class there is a function called SpawnRooms which spawns a r x c sized floor in which can be walked around in by the player (note: i didn't put any error handling on it, so don't break nothin by doing negatives n such).

The teleporting should work as long as the player object is tagged as Player