Modified the balance variables file, made a new game manager file, and modified BalanceMenu.cs
BalanceVariables.cs: made the variables into dictionaries and added more

BalanceMenu: now uses those dictionary variables, now accounts for a timer that is managed in the gamemanager. to modify values it now goes through the game manager

newGameManager.cs: made this to make sure their is no conflict if others are working on it. Now contains a timer for the balance, modified the way things are balanced